Шаблонизация и буферизация
